src/main/java/com/JavaWebSerice/WebService/utlis/JWTUtil.java
Incorporated JWT in utils package for session handling and authentication in UserDao, userDaoImp, authController & UserController

+++ src/main/java/com/JavaWebSerice/WebService/controllers/UserController.java
@@ -2,6 +2,7 @@ package com.JavaWebSerice.WebService.controllers;
 
 import com.JavaWebSerice.WebService.dao.UserDao;
 import com.JavaWebSerice.WebService.models.User;
+import com.JavaWebSerice.WebService.utlis.JWTUtil;
 import de.mkammerer.argon2.Argon2;
 import de.mkammerer.argon2.Argon2Factory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,6 +16,8 @@ public class UserController {
 
     @Autowired
     private UserDao userDao;
+    @Autowired
+    private JWTUtil jwtUtil;
 
     @RequestMapping(value = "api/users/{id}")
     public User getUser(@PathVariable Long id) {
@@ -27,9 +30,26 @@ public class UserController {
         user.setPhone("341555555");
         return user;
     }
+
+    //token validation method
+    private boolean validateToken(String token) {
+               /*
+        We must validate the token to allow requests (ie. see al users or delete)
+        We can select for key of value of jwtUtil. When we built the token we parsed 2 values (id &value)
+        in key it return id of user. We extract the user id.
+         */
+
+        String userId = jwtUtil.getKey(token);
+        return userId != null;
+
+    }
+
     //returns full list of users
     @RequestMapping(value = "api/users", method = RequestMethod.GET)
-    public List<User> listUsers() {
+    public List<User> listUsers(@RequestHeader(value="Authorization") String token) {
+        if (!validateToken(token)) {
+            return null;
+        }
         return userDao.listUsers();
     }
 
@@ -49,19 +69,12 @@ public class UserController {
     }
 
     @RequestMapping(value = "api/users/{id}", method = RequestMethod.DELETE)
-    public void deleteUser(@PathVariable long id) {
+    public void deleteUser(@RequestHeader(value="Authorization") String token, @PathVariable long id) {
+        //we verify in delete
+        if (!validateToken(token)) {
+            return;
+        }
         userDao.delete(id);
-    }
-
-    @RequestMapping(value = "userMod")
-    public User modUser() {
-        User user = new User();
-        user.setName("Faucndo");
-        user.setSurname("Bardi");
-        user.setEmail("facundo@mail.com");
-        user.setPassword("12341234");
-        user.setPhone("341555555");
-        return user;
     }
 
 }
